Multi-platform cache. References: #44
