+++ .github/workflows/check.yml
@@ -39,3 +39,15 @@ jobs:
       with:
         name: testDebugUnitTest
         path: ./**/build/reports/tests/testDebugUnitTest
+    - name: Upload test results
+      uses: actions/upload-artifact@v2
+      if: failure()
+      with:
+        name: tests
+        path: ./**/build/reports/tests/test
+    - name: Upload allTests results
+      uses: actions/upload-artifact@v2
+      if: failure()
+      with:
+        name: allTests
+        path: ./**/build/reports/tests/allTests
