+++ cache/src/commonTest/kotlin/com/motorro/rxlcemodel/cache/ClockMock.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2022 Nikolai Kotchetkov.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.motorro.rxlcemodel.cache
+
+import com.motorro.rxlcemodel.utils.Clock
+
+class ClockMock(private vararg val values: Long) : Clock {
+
+    private val iterator = values.iterator()
+
+    /**
+     * Current milliseconds value
+     */
+    override fun getMillis(): Long = if (iterator.hasNext()) {
+        iterator.nextLong()
+    } else {
+        values.last()
+    }
+}
